# Conflicts:
#	src/main/java/com/igushkin/homeworks/Main.java
#	src/main/java/com/igushkin/homeworks/lesson9/fileUtilities/FileUtilities.java
#	src/test/java/com/igushkin/homeworks/lesson9/fileUtilities/FileUtilitiesTest.java

+++ src/test/java/com/igushkin/homeworks/lesson9/fileUtilities/FileUtilitiesTest.java
@@ -11,8 +11,8 @@ class FileUtilitiesTest {
 
     @Test
     void countAnnotationsReturnsTwo() throws NotDirectoryException {
-        long expected = 2;
-        long actual = FileUtilities.countAnnotations(Main.DIRECTORY);
+        int expected = 2;
+        int actual = FileUtilities.countAnnotations(Main.DIRECTORY);
         assertEquals(actual, expected);
     }
